backend/requirements.txt
flask>=2.0.0
yt-dlp>=2021.12.1
ffmpeg-python>=0.2.0
werkzeug>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0